fix: fix reading info from google sheets
Fixed an error reading device information from Google sheets when the cell with device information is empty. Now the background image is not created even if at least one cell in the device information row is not filled in. This is reported to the user in the log.
The process of reading data from the application.properties file has also been optimized. Now the reading takes place once at the beginning of the program and there are no duplicate messages in the logs when generating each image.